requirements.txt
absl-py==1.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
